java-trader 基于 jdk17/gradle 7重构

+++ trader-common/src/main/java/trader/service/ta/AbsFutureBar.java
@@ -100,7 +100,7 @@ public abstract class AbsFutureBar implements FutureBar, JsonEnabled {
     }
 
     @Override
-    public int getTrades() {
+    public long getTrades() {
         return trades;
     }
